refactor : separate AddVideoDropdown Component
