+++ src/frontend/src/components/layouts/Header/BaseHeader.jsx
@@ -5,27 +5,21 @@ import { HeaderContext, MainLayoutContext } from '@utils/context';
 
 import { IconButton } from '@components/buttons';
 import {
-  Typography,
   HamburgerBar,
   AddEmptyCircle,
   AddFullCircle,
   FullAlarm,
   Alarm,
   Search,
-  MyVideo,
-  LiveStreaming,
 } from '@components/cores';
 import SearchForm from './SearchForm';
-import ProfileDropdown from './Dropdowns/ProfileDropdown';
+import AddVideoDropdown from './Dropdowns/AddVideoDropdown';
 import AlarmDropdown from './Dropdowns/AlarmDropdown';
+import ProfileDropdown from './Dropdowns/ProfileDropdown';
 
 function BaseHeader() {
   const { onToggle } = useContext(HeaderContext);
   const { onToggleSideNav, modalState, onToggleModal } = useContext(MainLayoutContext);
-
-  const handleAddVideoModalClick = e => {
-    e.stopPropagation();
-  };
 
   return (
     <>
@@ -44,20 +38,7 @@ function BaseHeader() {
           <IconButton name='addVideo' onClick={onToggleModal}>
             {modalState.addVideo ? <AddFullCircle /> : <AddEmptyCircle />}
           </IconButton>
-          <S.AddVideoDropdownContainer onClick={handleAddVideoModalClick}>
-            {modalState.addVideo && (
-              <S.AddVideoMenus>
-                <S.AddVideoMenu>
-                  <MyVideo />
-                  <Typography type='component'>동영상 업로드</Typography>
-                </S.AddVideoMenu>
-                <S.AddVideoMenu>
-                  <LiveStreaming />
-                  <Typography type='component'>실시간 스트리밍 시작</Typography>
-                </S.AddVideoMenu>
-              </S.AddVideoMenus>
-            )}
-          </S.AddVideoDropdownContainer>
+          <AddVideoDropdown />
         </div>
         <div>
           <IconButton name='alarm' onClick={onToggleModal}>
