+++ src/frontend/src/components/layouts/Header/Dropdowns/AddVideoDropdown.jsx
@@ -1,0 +1,34 @@
+import React, { useContext } from 'react';
+
+import * as S from './Dropdown.style';
+import { MainLayoutContext } from '@utils/context';
+import { modalService } from '@utils/service';
+
+import { Typography, MyVideo, LiveStreaming } from '@components/cores';
+
+function AddVideoDropdown() {
+  const { modalState } = useContext(MainLayoutContext);
+
+  const handleAddVideoModalClick = e => {
+    e.stopPropagation();
+  };
+
+  return (
+    <S.AddVideoDropdownContainer onClick={handleAddVideoModalClick}>
+      {modalState.addVideo && (
+        <S.AddVideoMenus>
+          <S.AddVideoMenu>
+            <MyVideo />
+            <Typography type='component'>동영상 업로드</Typography>
+          </S.AddVideoMenu>
+          <S.AddVideoMenu>
+            <LiveStreaming />
+            <Typography type='component'>실시간 스트리밍 시작</Typography>
+          </S.AddVideoMenu>
+        </S.AddVideoMenus>
+      )}
+    </S.AddVideoDropdownContainer>
+  );
+}
+
+export default AddVideoDropdown;
